Implement JUnit 5 Test Executor
Run each test separately and accumulate all the test result
together, which means invoking Launcher.execute for each test method.
So we can have the full control of the execution order.
Note that this is a workaround. Running tests separately causes some
overhead. A better solution is to run all tests with single
Launcher.execute.